Fix final callback to be called
The full async chain was not being fulfilled because glob.sync would include './../webdriverio/lib/commands' (a directory) as an array item when calling the process with `bin/wddoc -i ./../webdriverio/lib/commands/ -t ./templates/template.md.ejs`. The fix just corrects glob so that it only returns the array with filenames, not the directory path.